+++ changelog.md
@@ -1,1 +1,1 @@
-- Fixed Jade integration
+- Updated to 1.21.7